Create the YouTubeHandler class
- Move functions from the MasterSite class
